+++ lib/src/view.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_clean_architecture/src/controller.dart';
+import 'package:flutter_clean_architecture/flutter_clean_architecture.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:provider/provider.dart';
+import 'package:responsive_builder/responsive_builder.dart';
 
 enum ScreenSizeType {
   TABLET,
@@ -10,11 +11,14 @@ enum ScreenSizeType {
   MOBILE,
 }
 
+/// Defines a function type that receives a [BuildContext] and returns a [Widget] for widget builders.
+typedef ViewBuilder = Widget Function(BuildContext context);
+
 /// The [ResponsiveViewState] represents the [State] of a [StatefulWidget], typically of a screen or a
 /// page. The [ResponsiveViewState] requires a [Controller] to handle its events and provide its data.
 ///
-/// The [ResponsiveViewState] allow us to provide until three build methods to abstract responsivity for the
-/// developer, and the screen renders the view based on [MediaQuery] screen width.
+/// The [ResponsiveViewState] allow us to provide until four build methods to abstract responsiveness for the
+/// developer, and the screen renders the view based on [MediaQuery] screen width managed by [ScreenTypeLayout.builder].
 ///
 ///
 /// The [ResponsiveViewState] also has a default [globalKey] that can be used inside its `builds` function
@@ -22,110 +26,67 @@ enum ScreenSizeType {
 /// snackbars, dialogs, and so on.
 ///
 /// The [ResponsiveViewState] lifecycle is also handled by the [Controller].
+///
+/// You can optionally define builders for layouts that you want to implement (It will always give priority to bigger to smaller
+/// resolutions (e.g: If desktop isn't provided on desktop resolution, it will try to build tablet and forward)
 /// ```dart
-///     class CounterState extends ViewResponsiveState<CounterPage, CounterController> {
+///     class CounterState extends ResponsiveViewState<CounterPage, CounterController> {
 ///       CounterState(CounterController controller) : super(controller);
 ///
 ///       @override
-///       Widget buildMobileView(BuildContext context) {
+///       ViewBuilder mobileBuilder = (BuildContext context) {
 ///         return Text("Mobile view");
-///       }
+///       };
 ///
 ///       @override
-///       Widget buildTabletView(BuildContext context) {
+///       ViewBuilder tabletBuilder = (BuildContext context) {
 ///         return Text("Tablet view");
-///       }
+///       };
 ///
 ///       @override
-///       Widget buildDesktopBiew(BuildContext context) {
+///       ViewBuilder desktopBuilder = (BuildContext context) {
 ///         return Text("Desktop view");
-///       }
+///       };
+///
+///       @override
+///       ViewBuilder watchBuilder = (BuildContext context) {
+///         return Text("Watch view");
+///       };
 ///     }
 /// ```
+///
+/// You can optionally set globally new default values for breakpoints. To do so, just check on [FlutterCleanArchitecture.setDefaultViewBreakpoints]
 abstract class ResponsiveViewState<Page extends View, Con extends Controller>
     extends ViewState<Page, Con> {
-  /// To fill breakpoint params, they must be passed on super with it's name.
-  /// ```dart
-  /// SomePageState(SomeController controller)
-  /// : super(
-  ///     controller,
-  ///     tabletBreakpointMinimumWidth: 700,
-  ///     desktopBreakpointMinimumWidth: 1200,
-  ///   );
-  /// ```
-  ///
-  ResponsiveViewState(
-    Con controller, {
-    this.tabletBreakpointMinimumWidth = 600,
-    this.desktopBreakpointMinimumWidth = 1024,
-  })  : assert(desktopBreakpointMinimumWidth > tabletBreakpointMinimumWidth,
-            'Desktop breakpoint must not be less than tablet'),
-        super(controller);
+  ResponsiveViewState(Con controller) : super(controller);
 
-  /// This breakpoint targets the minimum width of [Tablet] size. The default value is 600.
-  /// When the width size from [context] comes under 600 (or the given value), it automatically switchs to [Mobile Viewport].
-  final double tabletBreakpointMinimumWidth;
+  /// Abstract builder to be implemented by the developer which will build on [Watch ViewPort].
+  ///   /// The default breakpoint value is less than [300]
+  ViewBuilder watchBuilder;
 
-  /// This breakpoint targets the minimum width of [Desktop] size. The default value is 1024.
-  /// When the width size from [context] comes under 1024 (or the given value), it automatically switchs to [Tablet Viewport].
-  final double desktopBreakpointMinimumWidth;
+  /// Abstract builder to be implemented by the developer which will build on [Mobile ViewPort].
+  /// The default breakpoint value is more than [300]
+  ViewBuilder mobileBuilder;
 
-  /// Abstract Method to be implemented by the developer which implements [Mobile ViewPort].
-  Widget buildMobileView();
+  /// Abstract builder to be implemented by the developer which will build on [Tablet/Pad ViewPort].
+  ///   /// The default breakpoint value is [600]
+  ViewBuilder tabletBuilder;
 
-  /// Abstract Method to be implemented by the developer which implements [Tablet/Pad ViewPort].
-  Widget buildTabletView();
+  /// Abstract builder to be implemented by the developer which will build on [Desktop ViewPort].
+  ///   /// The default breakpoint value is [950]
+  ViewBuilder desktopBuilder;
 
-  /// Abstract Method to be implemented by the developer which implements [Desktop ViewPort].
-  Widget buildDesktopView();
-
-  /// This method verify the dimensions using [MediaQuery], and so it defines which viewport will be exposed: [MOBILE], [TABLET] or [DESKTOP].
-  /// The Default ViewPort is [MOBILE].
-  ScreenSizeType get _screenSizeType {
-    final screenWidth = MediaQuery.of(context).size.width;
-
-    if (screenWidth < tabletBreakpointMinimumWidth) {
-      return ScreenSizeType.MOBILE;
-    }
-
-    if (screenWidth < desktopBreakpointMinimumWidth &&
-        screenWidth >= tabletBreakpointMinimumWidth) {
-      return ScreenSizeType.TABLET;
-    }
-
-    return ScreenSizeType.DESKTOP;
-  }
-
-  /// This turns buildPage into an implicit method that build according to the given builds methods: [MOBILE], [TABLET] and [DESKTOP].
+  /// This turns buildPage into an implicit method that build according to the given builds methods: [MOBILE], [TABLET], [DESKTOP] and [WATCH].
   /// The Default Viewport is [MOBILE]. When [TABLET] or [DESKTOP] builds are null, [MOBILE] viewport will be called.
+
   @override
   @nonVirtual
-  Widget buildPage() {
-    return LayoutBuilder(
-      builder: (BuildContext context, BoxConstraints constraints) {
-        switch (_screenSizeType) {
-
-          /// In case of screen width fits the mobile breakpoint, it should try to build the mobile view into the widget.
-          /// In case of the mobile view method isn't defined, it will try to render the tablet view, and the desktop view.
-          /// If any of these are defined, it will return null.
-          case ScreenSizeType.MOBILE:
-            return buildMobileView() ?? buildTabletView() ?? buildDesktopView();
-
-          /// In case of screen width fits the tablet breakpoint, it should try to build the tablet view into the widget.
-          /// In case of the tablet view method isn't defined, it will try to render the desktop view, and the mobile view.
-          /// If any of these are defined, it will return null.
-          case ScreenSizeType.TABLET:
-            return buildTabletView() ?? buildDesktopView() ?? buildMobileView();
-
-          /// In case of screen width fits the mobile breakpoint, it should try to build the desktop view into the widget.
-          /// In case of the mobile view method isn't defined, it will try to render the tablet view, and the mobile view.
-          /// If any of these are defined, it will return null.
-          case ScreenSizeType.DESKTOP:
-            return buildDesktopView() ?? buildTabletView() ?? buildMobileView();
-        }
-
-        return null;
-      },
+  Widget get view {
+    return ScreenTypeLayout.builder(
+      mobile: mobileBuilder,
+      tablet: tabletBuilder,
+      desktop: desktopBuilder,
+      watch: watchBuilder,
     );
   }
 }
@@ -170,9 +131,13 @@ abstract class ViewState<Page extends View, Con extends Controller>
     extends State<Page> {
   final GlobalKey<State<StatefulWidget>> globalKey =
       GlobalKey<State<StatefulWidget>>();
-  Con _controller;
+  final Con _controller;
   Logger _logger;
-  Con get controller => _controller;
+  ViewBuilder builder;
+
+  /// Implement the [Widget] you want to be displayed on [View]
+  Widget get view;
+
   ViewState(this._controller) {
     _controller.initController(globalKey);
     WidgetsBinding.instance.addObserver(_controller);
@@ -190,17 +155,10 @@ abstract class ViewState<Page extends View, Con extends Controller>
     super.didChangeDependencies();
   }
 
-  Widget buildPage();
-
   @override
   @nonVirtual
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider<Con>.value(
-        value: _controller,
-        child: Consumer<Con>(builder: (ctx, con, _) {
-          _controller = con;
-          return buildPage();
-        }));
+    return ChangeNotifierProvider<Con>.value(value: _controller, child: view);
   }
 
   @override