+++ test/controller_test.dart
@@ -7,14 +7,30 @@ GlobalKey snackBar = GlobalKey();
 GlobalKey inc = GlobalKey();
 
 void main() {
+  var numberOfWidgetBuilds = 0;
+  var numberOfUncontrolledWidgetBuilds = 0;
+  var numberOfControlledWidgetBuilds = 0;
+
   testWidgets('Controller can change data and refresh View',
       (WidgetTester tester) async {
     final AutomatedTestWidgetsFlutterBinding binding = tester.binding;
     binding.addTime(const Duration(seconds: 3));
-    await tester.pumpWidget(CounterPage());
+    await tester.pumpWidget(MaterialApp(
+      home: CounterPage(
+        onWidgetBuild: () {
+          numberOfWidgetBuilds++;
+        },
+        onUncontrolledWidgetBuild: () {
+          numberOfUncontrolledWidgetBuilds++;
+        },
+        onControlledWidgetBuild: () {
+          numberOfControlledWidgetBuilds++;
+        },
+      ),
+    ));
 
     // Create our Finders
-    Finder counterFinder = find.text('0');
+    var counterFinder = find.text('0');
     expect(counterFinder, findsOneWidget);
 
     await tester.tap(find.byKey(inc));
@@ -34,6 +50,10 @@ void main() {
     await tester.tap(find.byKey(snackBar));
     await tester.pump();
     expect(find.text('Hi'), findsOneWidget);
+
+    expect(numberOfWidgetBuilds, equals(1));
+    expect(numberOfUncontrolledWidgetBuilds, equals(1));
+    expect(numberOfControlledWidgetBuilds, equals(3));
   });
 }
 
@@ -58,6 +78,15 @@ class CounterController extends Controller {
 }
 
 class CounterPage extends View {
+  final Function onWidgetBuild;
+  final Function onUncontrolledWidgetBuild;
+  final Function onControlledWidgetBuild;
+
+  CounterPage(
+      {this.onWidgetBuild,
+      this.onUncontrolledWidgetBuild,
+      this.onControlledWidgetBuild});
+
   @override
   State<StatefulWidget> createState() => CounterState();
 }
@@ -66,21 +95,44 @@ class CounterState extends ViewState<CounterPage, CounterController> {
   CounterState() : super(CounterController());
 
   @override
-  Widget buildPage() {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      home: Scaffold(
-        key: globalKey,
-        body: Column(
-          children: <Widget>[
-            Center(
-              child: Text(controller.counter.toString()),
+  Widget get view {
+    widget.onWidgetBuild();
+
+    return Scaffold(
+      key: globalKey,
+      body: Column(
+        children: <Widget>[
+          Center(
+            child: Builder(
+              builder: (BuildContext context) {
+                widget.onUncontrolledWidgetBuild();
+
+                return Text('Uncontrolled text');
+              },
             ),
-            MaterialButton(key: inc, onPressed: () => controller.increment()),
-            MaterialButton(
-                key: snackBar, onPressed: () => controller.showSnackBar()),
-          ],
-        ),
+          ),
+          Center(
+            child: ControlledWidgetBuilder<CounterController>(
+              builder: (ctx, controller) {
+                widget.onControlledWidgetBuild();
+
+                return Text(controller.counter.toString());
+              },
+            ),
+          ),
+          ControlledWidgetBuilder<CounterController>(
+            builder: (ctx, controller) {
+              return MaterialButton(
+                  key: inc, onPressed: () => controller.increment());
+            },
+          ),
+          ControlledWidgetBuilder<CounterController>(
+            builder: (ctx, controller) {
+              return MaterialButton(
+                  key: snackBar, onPressed: () => controller.showSnackBar());
+            },
+          ),
+        ],
       ),
     );
   }