+++ CHANGELOG.md
@@ -1,5 +1,27 @@
-## [3.1.0] - Monday, June 8th, 2020
+## [4.0.0] - Thursday, September 24st, 2020
+### What's New
+- Created `ControlledWidget` to create `StatelessWidgets` refreshable by `Controller`
+    - Now, to create refreshable widgets, use `ControlledWidget` builder.
+    - When `Controller.refreshUI` is called, only `ControlledWidgets` will re-render
+    - To create controlled `StatefulWidgets`, please check `FlutterCleanArchitecture.getController` approach
+- Improves `ResponsiveViewState` with watch interface
+- Added `FlutterCleanArchitecture.setDefaultViewBreakpoints` to configure view breakpoints globally
+- Added `mobileBuilder` to `ResponsiveViewState`
+- Added `tabletBuilder` to `ResponsiveViewState`
+- Added `desktopBuilder` to `ResponsiveViewState`
+- Added `watchBuilder` to `ResponsiveViewState`
 
+### Breaking Changes
+- Removed getter `controller` from `ViewState`
+- Removed `buildTabletView()`
+- Removed `buildMobileView()`
+- Removed `buildDesktopView()`
+
+## [3.1.1] - Monday, September 21st, 2020
+
+- Added an option to pass listen:false when calling getController() outside of the build method.
+
+## [3.1.0] - Monday, June 8th, 2020
 - Created Github Actions to run analyze and tests on package
 - Created responsive view to improve the usage on flutter web
 - Fixed lint rules to fit pedantic 1.9.0